reorganized project files a little bit, got rid of components that were not being used
