--- src/components/AboutPage/AboutPage.js
@@ -1,16 +1,0 @@
-import React from 'react';
-
-// This is one of our simplest components
-// It doesn't have local state, so it can be a function component.
-// It doesn't dispatch any redux actions or display any part of redux state
-// or even care what the redux state is, so it doesn't need 'connect()'
-
-const AboutPage = () => (
-  <div className="container">
-    <div>
-      <h1>About Page</h1>
-    </div>
-  </div>
-);
-
-export default AboutPage;
