+++ src/components/Filter/FilterOptions/MakerOptions.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { connect } from 'react-redux';
-import mapStoreToProps from '../../redux/mapStoreToProps';
+import mapStoreToProps from '../../../redux/mapStoreToProps';
 import { Collapse, Button, Form, FormGroup, Label, Input } from 'reactstrap';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faChevronUp, faChevronDown } from '@fortawesome/free-solid-svg-icons'
@@ -8,16 +8,16 @@ import { faChevronUp, faChevronDown } from '@fortawesome/free-solid-svg-icons'
 const upArrow = <FontAwesomeIcon icon={faChevronUp}/>
 const downArrow = <FontAwesomeIcon icon={faChevronDown}/>
 
-const LocationOptions = (props) => {
+const MakerOptions = (props) => {
   const [isOpen, setIsOpen] = useState(false);
 
   const toggle = () => setIsOpen(!isOpen);
 
   const handleChange = (event) => {
     if (document.getElementById(event.target.id).checked) {
-        props.dispatch({type: 'ADD_LOCATION', payload: event.target.id});
+        props.dispatch({type: 'ADD_MAKER_OPTIONS', payload: event.target.id});
     } else {
-        props.dispatch({type: 'REMOVE_LOCATION', payload: event.target.id});
+        props.dispatch({type: 'REMOVE_MAKER_OPTIONS', payload: event.target.id});
     }
   }
 
@@ -27,85 +27,98 @@ const LocationOptions = (props) => {
             className="filterBtn"
             onClick={toggle}
             style={{ marginBottom: '1rem' }}>
-                Location Options {isOpen ? <span className="dots">{upArrow}</span> : <span className="dots">{downArrow}</span>}
+                Maker Types {isOpen ? <span className="dots">{upArrow}</span> : <span className="dots">{downArrow}</span>}
         </Button>
 
         <Collapse isOpen={isOpen} className="collapseContainer">
             <Form>
                 <FormGroup check>
                     <Label check>
-                        <Input type="checkbox" id="grocery" onClick={(event) => handleChange(event)} />{' '}
-                            Local Grocery Stores
+                        <Input type="checkbox" id="farmer" onClick={(event) => handleChange(event)} />{' '}
+                            Farmer/Grower
                     </Label>
                 </FormGroup>
                 
                 <FormGroup check>
                     <Label check>
-                        <Input type="checkbox" id="co_ops" onClick={(event) => handleChange(event)} />{' '}
-                            Co-ops
+                        <Input type="checkbox" id="forager" onClick={(event) => handleChange(event)} />{' '}
+                            Forager
                     </Label>
                 </FormGroup>
                 
                 <FormGroup check>
                     <Label check>
-                        <Input type="checkbox" id="farmers_market" onClick={(event) => handleChange(event)} />{' '}
-                            Farmers Markets/Other Local Markets
+                        <Input type="checkbox" id="maker" onClick={(event) => handleChange(event)} />{' '}
+                            Maker
                     </Label>
                 </FormGroup>
                 
                 <FormGroup check>
                     <Label check>
-                        <Input type="checkbox" id="amazon" onClick={(event) => handleChange(event)} />{' '}
-                        Online on Amazon
+                        <Input type="checkbox" id="distiller" onClick={(event) => handleChange(event)} />{' '}
+                            Distiller
                     </Label>
                 </FormGroup>
                 
                 <FormGroup check>
                     <Label check>
-                        <Input type="checkbox" id="company_website" onClick={(event) => handleChange(event)} />{' '}
-                        Online on Company Website
+                        <Input type="checkbox" id="brewer" onClick={(event) => handleChange(event)} />{' '}
+                            Brewer
                     </Label>
                 </FormGroup>
                 
                 <FormGroup check>
                     <Label check>
-                        <Input type="checkbox" id="gift" onClick={(event) => handleChange(event)} />{' '}
-                        Gift/Subscription Box
+                        <Input type="checkbox" id="winemaker" onClick={(event) => handleChange(event)} />{' '}
+                            Wine Maker
                     </Label>
                 </FormGroup>
                 
                 <FormGroup check>
                     <Label check>
-                        <Input type="checkbox" id="retail" onClick={(event) => handleChange(event)} />{' '}
-                        Retail Store (Non-Grocery)
+                        <Input type="checkbox" id="cidery" onClick={(event) => handleChange(event)} />{' '}
+                            Cidery
                     </Label>
                 </FormGroup>
                 
                 <FormGroup check>
                     <Label check>
-                        <Input type="checkbox" id="your_home" onClick={(event) => handleChange(event)} />{' '}
-                        Private Place of Business, Farm, or Home
+                        <Input type="checkbox" id="roaster" onClick={(event) => handleChange(event)} />{' '}
+                            Roaster
                     </Label>
                 </FormGroup>
                 
                 <FormGroup check>
                     <Label check>
-                        <Input type="checkbox" id="roadside" onClick={(event) => handleChange(event)} />{' '}
-                        Roadside Stand
+                        <Input type="checkbox" id="apiary" onClick={(event) => handleChange(event)} />{' '}
+                            Apiary (Bees)
+                    </Label>
+                </FormGroup>
+                
+                <FormGroup check>
+                    <Label check>
+                        <Input type="checkbox" id="woman_owned" onClick={(event) => handleChange(event)} />{' '}
+                            Woman Owned
+                    </Label>
+                </FormGroup>
+                
+                <FormGroup check>
+                    <Label check>
+                        <Input type="checkbox" id="poc_owned" onClick={(event) => handleChange(event)} />{' '}
+                            P.O.C. Owned
                     </Label>
                 </FormGroup>
                 
                 <FormGroup check>
                     <Label check>
                         <Input type="checkbox" id="other" onClick={(event) => handleChange(event)} />{' '}
-                        Other
+                            Other
                     </Label>
                 </FormGroup>
-
             </Form>
         </Collapse>
     </div>
   );
 }
 
-export default connect(mapStoreToProps)(LocationOptions);
+export default connect(mapStoreToProps)(MakerOptions);
