+++ src/components/Login/LoginForm/LoginForm.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import mapStoreToProps from '../../redux/mapStoreToProps';
+import mapStoreToProps from '../../../redux/mapStoreToProps';
 import {withRouter} from 'react-router-dom';
 import { Button, Form, FormGroup, Label, Input } from 'reactstrap';
