+++ src/components/Profiles/ProfilePage/ProfilePage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 // import { Link } from 'react-router-dom';
-import mapStoreToProps from '../../redux/mapStoreToProps';
+import mapStoreToProps from '../../../redux/mapStoreToProps';
 import MakerProfile from '../MakerProfile/MakerProfile';
 import UserProfile from '../UserPage/UserPage';
 
@@ -14,11 +14,12 @@ class ProfilePage extends React.Component {
   }
   render() {
     return (
-
-      
         <>
+        {this.props.store.maker.id ?
         <MakerProfile/>
+        :
         <UserProfile/>
+        }
         </>
 
     )
